[feat, fix] 함수및 변수명 정리 및 랭크게임 일반 게임에 따른 분기 설정 #405 #406
