+++ components/modal/afterGame/score/Buttons.tsx
