+++ components/modal/afterGame/score/Guide.tsx
