+++ components/modal/afterGame/score/Score.tsx
