+++ components/modal/match/MatchManualModal.tsx
@@ -3,10 +3,10 @@ import { modalState } from 'utils/recoil/modal';
 import styles from 'styles/modal/MatchManualModal.module.scss';
 
 export default function MatchManualModal() {
-  const setModalInfo = useSetRecoilState(modalState);
+  const setModal = useSetRecoilState(modalState);
 
   const onReturn = () => {
-    setModalInfo({ modalName: null });
+    setModal({ modalName: null });
   };
 
   return (
